osai-lab/targets/mcp/server/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
psycopg2-binary==2.9.10
